+++ client/components/Treemap.jsx
@@ -1,5 +1,5 @@
 import {Component} from 'preact';
-import FoamTree from 'carrotsearch.foamtree';
+import FoamTree from '@carrotsearch/foamtree';
 
 export default class Treemap extends Component {