+++ CHANGELOG.md
@@ -20,6 +20,10 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
   * Don't show an empty tooltip when hovering over the FoamTree attribution group or between top-level groups ([#413](https://github.com/webpack-contrib/webpack-bundle-analyzer/pull/413) by [@
     stanislawosinski](https://github.com/stanislawosinski))
 
+* **Internal**
+  * Upgrade FoamTree to version 3.5.0, replace vendor dependency with an NPM package ([#412](https://github.com/webpack-contrib/webpack-bundle-analyzer/pull/412) by [@
+    stanislawosinski](https://github.com/stanislawosinski))
+    
 ## 4.3.0
 
  * **Improvement**